edit and delete done

+++ app/profile/page.tsx
@@ -3,16 +3,33 @@ import React, { useEffect, useState } from "react";
 import Profile from "@components/Profile";
 import { useSession } from "next-auth/react";
 import { IPrompt } from "@customTypes/prompt";
+import { useRouter } from "next/navigation";
+
 const ProfilePage = () => {
+  const router = useRouter();
   const { data: session } = useSession();
-  const handleEdit = () => {};
-  const handleDelete = async (promptId: string) => {
-    const response = await fetch(`/api/prompt/${promptId}/`, {
-      method: "DELETE",
-    });
-    const data = await response.json();
+  const handleEdit = (post: IPrompt) => {
+    router.push(`update-prompt?id=${post._id}`);
   };
   const [posts, setPosts] = useState<IPrompt[]>([]);
+  const handleDelete = async (promptId: string) => {
+    const isDeleteConfirmed = confirm("Are you sure want to Delele");
+    if (isDeleteConfirmed) {
+      try {
+        const response = await fetch(`/api/prompt/${promptId}/`, {
+          method: "DELETE",
+        });
+        const data = (await response.json()) as IPrompt;
+        const filteredPosts: IPrompt[] = posts.filter(
+          (x) => x._id !== data._id
+        );
+        setPosts(filteredPosts);
+      } catch (error) {
+        console.log(error);
+      }
+    }
+  };
+
   useEffect(() => {
     const fetchPosts = async () => {
       const response = await fetch(`/api/users/${session?.user.id}/posts`);
